+++ Runtime/Storage/FileSystem.cs
@@ -1,7 +1,7 @@
+using UnityEngine;
 using System.IO;
 using System.Diagnostics;
 using System.Threading.Tasks;
-using UnityEngine;
 
 namespace ActionCode.Persistence
 {
@@ -30,26 +30,21 @@ namespace ActionCode.Persistence
             cryptographer = CryptographerFactory.Create(cryptographerType, cryptographerKey);
         }
 
-        public async Task Save<T>(T data, string name)
+        public async Task Save<T>(T data, string name, bool saveRawData)
         {
             var path = GetPath(name);
             var content = serializer.Serialize(data);
+
+            if (saveRawData)
+            {
+                var rawPath = Path.ChangeExtension(path, serializer.Extension);
+                await SynchronyStreamAdapter.Write(rawPath, content);
+            }
 
             if (cryptographer != null) content = await cryptographer.Encrypt(content);
             if (compressor != null) content = await compressor.Compress(content);
 
             await SynchronyStreamAdapter.Write(path, content);
-        }
-
-        public void SaveUncompressed<T>(T data, string name)
-        {
-            var path = GetPath(name);
-            var content = serializer.Serialize(data);
-
-            path = Path.ChangeExtension(path, serializer.Extension);
-
-            using var writer = new StreamWriter(path);
-            writer.Write(content);
         }
 
         public bool TryLoad<T>(string name, out T data)